Updating to build from source
The quadratic outlines are quite simple and were easy to convert to cubic. There is an opportunity for optimization as there are an excess number of points, but wanted to do a first check in that just has the cubics and is building from source using the makefile.

Also note that there has been a change to the glyph names as the vertical-specific glyphs were previously named unhelpfully. When diffing you'll see that a number of glyphs are now "missing" due to this name change.

There is a slight increase in file size as a result of this update.